+++ inc/commoninjectionlib.class.php
@@ -1098,6 +1098,14 @@ class PluginDatainjectionCommonInjectionLib
                 //Get search option associated with the field
                 $option = self::findSearchOption($searchOptions, $field);
 
+                if ($option && !isset($option['checktype'])) {
+                    // In some cases, float datatype is wrongly detected as string, decimal or number, so we check that first.
+                    //Regex matches all decimal formats accepted by glpi
+                    $option['checktype'] = (preg_match($this->getFloatDetectionRegex(), $value) !== 0)
+                        ? "float"
+                        : $option['datatype'];
+                }
+
                 // Check some types
                 switch ($option['checktype'] ?? 'text') {
                     case "date":
@@ -1299,6 +1307,13 @@ class PluginDatainjectionCommonInjectionLib
         return $mac;
     }
 
+    //--------------------------------------------------//
+    //----------- Utility methods -----------------------//
+    //------------------------------------------------//
+    public function getFloatDetectionRegex(): string
+    {
+        return '/^(?:(?:\d{1,3}(?: \d{3})+|\d+)[.,]|\d{1,3}(?:,\d{3})+\.)\d+$/';
+    }
 
     //--------------------------------------------------//
     //----------- Check methods -----------------------//
@@ -1397,7 +1412,7 @@ class PluginDatainjectionCommonInjectionLib
     {
 
         if (!empty($option)) {
-            $field_type = ($option['checktype'] ?? 'text');
+            $field_type = ($option['checktype'] ?? $option['datatype'] ?? 'text');
 
             //If no data provided AND this mapping is not mandatory
             if (